refactor: fold control api into core api
