--- control_api_test.go
@@ -1,44 +1,0 @@
-package main
-
-import (
-	"context"
-	"net/http"
-	"net/http/httptest"
-	"testing"
-	"time"
-)
-
-func TestServeControlAPIShutsDown(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	done := make(chan struct{})
-	go func() {
-		serveControlAPI(ctx, Config{ControlAPIAddr: "127.0.0.1:0"}, nil)
-		close(done)
-	}()
-
-	select {
-	case <-done:
-	case <-time.After(time.Second):
-		t.Fatal("serveControlAPI did not return after context cancellation")
-	}
-}
-
-func TestControlAPIHealthzResponseShape(t *testing.T) {
-	mux := newControlMux(nil)
-
-	req, err := http.NewRequest(http.MethodGet, "/healthz", nil)
-	if err != nil {
-		t.Fatalf("new request: %v", err)
-	}
-	rec := httptest.NewRecorder()
-	mux.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusOK {
-		t.Fatalf("status = %d, want %d", rec.Code, http.StatusOK)
-	}
-	if rec.Body.String() != `{"status":"ok"}` {
-		t.Fatalf("body = %q, want health payload", rec.Body.String())
-	}
-}
